+++ src/address/mod.rs
@@ -211,7 +211,7 @@ where
     M: Message + Send,
     M::Result: Send,
 {
-    tx: Box<Sender<M>>,
+    tx: Box<dyn Sender<M>>,
 }
 
 impl<M> Recipient<M>
@@ -220,7 +220,7 @@ where
     M::Result: Send,
 {
     /// Creates a new recipient.
-    pub(crate) fn new(tx: Box<Sender<M>>) -> Recipient<M> {
+    pub(crate) fn new(tx: Box<dyn Sender<M>>) -> Recipient<M> {
         Recipient { tx }
     }
 
@@ -307,6 +307,16 @@ where
 {
     fn hash<H: Hasher>(&self, state: &mut H) {
         self.tx.hash().hash(state)
+    }
+}
+
+impl<M> fmt::Debug for Recipient<M>
+where
+    M: Message + Send,
+    M::Result: Send,
+{
+    fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
+        write!(fmt, "Recipient {{ /* omitted */ }}")
     }
 }