+++ src/context.rs
@@ -144,6 +144,11 @@ where
     pub fn set_mailbox_capacity(&mut self, cap: usize) {
         self.parts.set_mailbox_capacity(cap)
     }
+
+    /// Returns whether any addresses are still connected.
+    pub fn connected(&self) -> bool {
+        self.parts.connected()
+    }
 }
 
 impl<A> AsyncContextParts<A> for Context<A>