+++ CHANGES.md
@@ -1,5 +1,15 @@
 # CHANGES
 
+## [0.8.4] (unreleased)
+
+### Added
+
+* Added `Context::connected()` to check any addresses are alive
+
+### Changed
+
+* Remove `hashbrown` crate from dependencies
+
 ## [0.8.3] (2019-05-29)
 
 ### Fixed